requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
passlib[bcrypt]
python-dotenv

# testing
pytest
httpx
